+++ lib/top_secret/text/result.rb
@@ -3,7 +3,7 @@
 module TopSecret
   class Text
     # Holds the result of a redaction operation.
-    class Result
+    class Result # TODO: Rename to FilterResult
       # @return [String] The original unredacted input
       attr_reader :input