+++ lib/top_secret/text/scan_result.rb
@@ -4,17 +4,14 @@ module TopSecret
   class Text
     # Holds the result of a scan operation.
     class ScanResult
+      include Mapping
+
       # @return [Hash] Mapping of redacted labels to matched values
       attr_reader :mapping
 
       # @param mapping [Hash] Map of labels to matched values
       def initialize(mapping)
         @mapping = mapping
-      end
-
-      # @return [Boolean] Whether sensitive information was found
-      def sensitive?
-        mapping.any?
       end
     end
   end